live.py is fixed
The portfolio_strategy.py used to rise an error saying pnl_bok and positions_book are not readable. which means when the bot was running for the first time the strategy.py couldn't comprehend that empty tables means whatever quantity of ticker in positions   table you are looking for is 0.  This is handled by having a condition if the tables are empty insert the quantity 0 for that ticker.
The previous folder <MQSMASTER\data_infra\tradingOps\backtest> is deleted.