NLP  & CI/CD Pipeline initialization
* Sample Bot

* Updated Visualise_backtest.ipynb

Gets news from FMP, calculates the sentiment score and gives the plot with stock price and sentiment score.

* Update visualize_NLP.ipynb

Added forecasting and sentiment score.

* update for NLP

run the contents of the file in colab by changing the runtime to T4 GPU

* Update visualize_NLP.ipynb

LSTM Model used for forecasting. The model predicts next-day stock prices using a 5-day lookback period. Further adjustments in progress.

* Update visualize_NLP.ipynb

LSTM model updated with 7 day forecast with sentiment integration.

* fetch_articles.py created

Created the fetch_articles.py which extracts the news and stores them in csv in MQSMASTER/NLP/articles folder.

* logging time

* EDITS

* os.getenv

* raw_sentiment vs change_in_price

* fetch_articles.py

* CAPM used

* Updated the `visualize_backtests.ipynb` file to exclude weekenda from the graphs.

* CAPM updated

* Fetch_articles.py updated

* Updated Graphs

fixed the issue in MQSDBconnector which was taking the system's username instead of username from the .env file and updated the NLP graphs to plotly

* cell to save the model

* fixed fetch articles

* Created README.md for NLP folder

Included a README file which also includes a link for all the different models for easier accessibility.

* Updated Readme.md for NLP

* Updated the ploty graphs

* Plotly updated for adjustable date ranges

* In Visualise_NLP Correlation is plotted as area in the background

* Assest allocation

Scripts for calculating risk ratio were created.

* Revert "Assest allocation"

This reverts commit f1531496d16f01f9dfe3da1a7c74452cfd83921a.

* Asset Allocation Risk scripts

Created and updated the scripts that analysis risk.

* Update .gitignore

minor adjustments

* Update requirements.txt with corrected package versions

* Set up CI workflow

* Updated the yml file to Use GitHub secrets for DB credentials in CI

* ci: Add verbose logging to pytest

* ci: Added conftest.py

* Updated CI and created setup.py

* Updated yml file with the correct path to the tests

* Fixed graph error

* Update backtest_engine.py

---------

Co-authored-by: Simanta <simanta2000green@gmail.com>
Co-authored-by: ogadegun404 <gideonadegun@gmail.com>
Co-authored-by: mpundu29 <mkchikoya@mun.ca>
Co-authored-by: Fariha Mehnaz <fmehnaz@mun.ca>